* callocator
* cs; static local functions prefix with s_

+++ source/main/include/callocator/c_allocator_cs.h
@@ -24,7 +24,7 @@ namespace ncore
         {
             inline allocator_t() : m_object(nullptr) {}
 
-            bool setup(alloc_t* allocator, u16 max_object_instances, u16 max_components, u16 max_tags);
+            bool setup(alloc_t* allocator, u32 max_object_instances, u16 max_components, u8 max_tags);
             void teardown();
 
             // --------------------------------------------------------------------------------------------------
@@ -67,9 +67,9 @@ namespace ncore
             template <typename C1, typename C2> C2 const* get_component(C1 const* cp1) const { return (C2*)get_cp(C1::__ncs_index__, cp1, C2::__ncs_index__); }
 
             // Tags
-            template <typename C> bool has_tag(C const* cp, u16 tg_index) const { return has_tag(C::__ncs_index__, cp, tg_index); }
-            template <typename C> void add_tag(C const* cp, u16 tg_index) { add_tag(C::__ncs_index__, cp, tg_index); }
-            template <typename C> void rem_tag(C const* cp, u16 tg_index) { rem_tag(C::__ncs_index__, cp, tg_index); }
+            template <typename C> bool has_tag(C const* cp, u8 tg_index) const { return has_tag(C::__ncs_index__, cp, tg_index); }
+            template <typename C> void add_tag(C const* cp, u8 tg_index) { add_tag(C::__ncs_index__, cp, tg_index); }
+            template <typename C> void rem_tag(C const* cp, u8 tg_index) { rem_tag(C::__ncs_index__, cp, tg_index); }
 
             struct object_t;
 
@@ -83,7 +83,7 @@ namespace ncore
 
             void* create_instance(u16 cp_index);
             void  destroy_instance(u16 cp_index, void* cp);
-            u16   get_number_of_instances(u16 cp_index) const;
+            u32   get_number_of_instances(u16 cp_index) const;
 
             void* iterate_begin(u16 cp_index) const;
             void* iterate_next(u16 cp1_index, void const* cp1_ptr) const;
@@ -94,9 +94,9 @@ namespace ncore
             void*       get_cp(u16 cp1_index, void* cp1_ptr, u16 cp_index);
             void const* get_cp(u16 cp1_index, void const* cp1_ptr, u16 cp_index) const;
 
-            bool has_tag(u16 cp1_index, void const* cp1_ptr, u16 tg_index) const;
-            void add_tag(u16 cp1_index, void const* cp1_ptr, u16 tg_index);
-            void rem_tag(u16 cp1_index, void const* cp1_ptr, u16 tg_index);
+            bool has_tag(u16 cp1_index, void const* cp1_ptr, u8 tg_index) const;
+            void add_tag(u16 cp1_index, void const* cp1_ptr, u8 tg_index);
+            void rem_tag(u16 cp1_index, void const* cp1_ptr, u8 tg_index);
         };
     } // namespace ncs
 } // namespace ncore
